api|refact: add CheckHuskyRequirements into util package
